requirements.txt
typer>=0.12.3
requests>=2.32.3
rich>=13.7.1
python-dotenv>=1.0.1
streamlit>=1.37.0
pytest>=8.3.0
